PERA-1550 Move sdk interfaces into sdk package
