--- common-sdk/src/main/kotlin/com/algorand/wallet/algosdk/AlgoAccountSdk.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright 2022 Pera Wallet, LDA
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License
- */
-
-package com.algorand.wallet.algosdk
-
-import com.algorand.wallet.algosdk.model.Algo25Account
-import com.algorand.wallet.algosdk.model.HdAccount
-
-interface AlgoAccountSdk {
-
-    fun createHdAccount(): HdAccount
-
-    fun recoverHdAccount(mnemonic: String): HdAccount?
-
-    fun createAlgo25Account(): Algo25Account
-
-    fun recoverAlgo25Account(mnemonic: String): Algo25Account?
-}
