+++ common-sdk/src/main/kotlin/com/algorand/wallet/algosdk/transaction/sdk/AlgoAccountSdk.kt
@@ -10,8 +10,18 @@
  * limitations under the License
  */
 
-package com.algorand.wallet.algosdk
+package com.algorand.wallet.algosdk.transaction.sdk
 
-interface AlgoSdkUtils {
-    fun isValidAddress(address: String?): Boolean
+import com.algorand.wallet.algosdk.model.Algo25Account
+import com.algorand.wallet.algosdk.model.HdAccount
+
+interface AlgoAccountSdk {
+
+    fun createHdAccount(): HdAccount
+
+    fun recoverHdAccount(mnemonic: String): HdAccount?
+
+    fun createAlgo25Account(): Algo25Account
+
+    fun recoverAlgo25Account(mnemonic: String): Algo25Account?
 }
