+++ common-sdk/src/main/kotlin/com/algorand/wallet/algosdk/transaction/sdk/AlgoAccountSdkImpl.kt
@@ -10,7 +10,7 @@
  * limitations under the License
  */
 
-package com.algorand.wallet.algosdk
+package com.algorand.wallet.algosdk.transaction.sdk
 
 import cash.z.ecc.android.bip39.Mnemonics
 import cash.z.ecc.android.bip39.toSeed
