+++ common-sdk/src/main/kotlin/com/algorand/wallet/deeplink/parser/query/AccountAddressQueryParser.kt
@@ -12,13 +12,13 @@
 
 package com.algorand.wallet.deeplink.parser.query
 
-import com.algorand.wallet.algosdk.AlgoSdkUtils
+import com.algorand.wallet.algosdk.transaction.sdk.AlgoSdkAddress
 import com.algorand.wallet.deeplink.model.PeraUri
 import com.algorand.wallet.deeplink.utils.isCoinbaseDeepLink
 import javax.inject.Inject
 
 internal class AccountAddressQueryParser @Inject constructor(
-    private val algoSdkUtils: AlgoSdkUtils
+    private val algoSdkAddress: AlgoSdkAddress
 ) : DeepLinkQueryParser<String?> {
 
     override fun parseQuery(peraUri: PeraUri): String? {
@@ -26,13 +26,13 @@ internal class AccountAddressQueryParser @Inject constructor(
             peraUri.isAppLink() -> getAddressFromAppLink(peraUri)
             isCoinbaseDeepLink(peraUri) -> getAccountAddressForCoinbase(peraUri)
             else -> peraUri.getQueryParam(ACCOUNT_ID_QUERY_KEY) ?: peraUri.host
-        }?.takeIf { algoSdkUtils.isValidAddress(it) } ?: peraUri.rawUri.takeIf { algoSdkUtils.isValidAddress(it) }
+        }?.takeIf { algoSdkAddress.isValid(it) } ?: peraUri.rawUri.takeIf { algoSdkAddress.isValid(it) }
     }
 
     private fun getAddressFromAppLink(uri: PeraUri): String? {
         return uri.path
             ?.split("/")
-            ?.firstOrNull { algoSdkUtils.isValidAddress(it) }
+            ?.firstOrNull { algoSdkAddress.isValid(it) }
     }
 
     private fun getAccountAddressForCoinbase(uri: PeraUri): String? {
